Allow for incomplete custom types
Specifically, I'd like to be able to get away with a type definition consisting in just the `to` property, like this:
`
    int4: {
      to: 23
    }
`
That's because I mereley want to use a name for a Postgres type OID and the usual conversions are already defined in this file. As a fallback, the default serialization in src/connection.js at line 912 is just fine.