pass servername to tls.connect
This addresses the inability to connect using SNI to a TCP proxy hosting multiple postgres instances behind itself. This is for example the case of neon.tech. With this patch it is possible to establish connection to neon compute-node with specifying `postgres('database_url', { ssl: true })`. Without this patch, the servername is not passed.

See https://nodejs.org/docs/latest/api/tls.html#tls_tls_connect_options_callback for details on servername and SNI extenstion to TLS.