Update README.md (Transactions added missing returning *)
Check issue: https://github.com/porsager/postgres/issues/649

This is a minor modification but debugging this has taken a couple of hours for me as I am slightly new to SQL syntax and to postgreSQL in general. I was trying to use the empty return from sql.begin but it turned out that the callback in sql.begin was the one returning the empty array even though the insert was successful.