fix: Remove array support from inferType.
I believe this was always broken, because the oid of a boolean[]
is not the same as the oid of a boolean, and AFAICT inferType
was resolving "boolean[] --> the oid of a boolean" without any
attempt to translate over to the `boolean[]` oid.

There's talk of removing `inferType` entirely, which would have to
wait until v4, but I think if this array-based support has always
been broken (afaiu?), we can just remove it immediately in a v3
bug fix release.

Fixes #471